Allow creating unnamed server workers
Creating named workers is useful but falls down if I want to create a
large group of workers to work with (for instance) a poolboy pool.

This is a very naive attempt at changing as little code as possible to
introduce an atom `:no_name` which will cause the
`KafkaEx.Server.start_link` method to not register any name for the
started process.