Remove default values for offset commit/fetch reqs
Also added guard clauses and matching conditions to ensure that the
values are set sanely, and updated the tests accordingly.  In one case,
I found that the test for OffsetFetch was actually passing in an
OffsetCommit request and the test was passing because the structs have
similar fields.

I suspect that this will break integration tests; I will fix those in
the next commit.