Allow disabling the default worker, update README
The motivation to disable the default worker is described in #68

I updated the README to point to config/config.exs for the application
env, and updated config/config.exs with comments describing how to use
each of the supported env settings.  I think this makes it easier to
keep everything up-to-date.