Add consumer_group reader for introspection
In an application with potentially many workers, this can help with
consistency as other code may need to know the consumer group.  One
alternative is to pass around the name of the consumer group wherever we
pass around the worker name/pid, but this is more error prone than being
able to introspect the given process to determine its consumer group.