Use node_id for broker comparisons
This removes potential confusion when the configured broker hostport is different
from the advertised one. Also, a bootstrap metadata call is added that quickly
replaces whatever was configured with the actual cluster state to make sure that
these comparisons always work as we now work with returned broker information
including node ids instead of just host:port from the bootstrap broker list
in the configuration.