Have GenConsumer handle message sets
i.e., instead of `handle_message/2` we have `handle_message_set/2`

It's very easy for the user to iterate over the batch, and there is at
least one use case for passing batches to the consumer: parallelized
processing of messages.

At Simpli.fi we have some applications where each message we receive
from Kafka some amount of work that can be done asynchronously, and
we've found a big performance increase by processing those in parallel.
There are some delivery guarantee implications to that, but it should be
up to the user to decide how to handle that.