Emit meaningful error message on join group error
There are a number of errors that can be returned:
https://cwiki.apache.org/confluence/display/KAFKA/A+Guide+To+The+Kafka+Protocol#AGuideToTheKafkaProtocol-JoinGroupResponse

Some of them may be recoverable, but the Erlang Way here seems like
crashing in order to avoid having to manage an error state.  The code
was crashing before, but it would crash with a MatchError which was not
terribly helpful.