More gracefully handle errors on leaving consumer group
This seems to be related to #247 though it may not be the root cause.

We were seeing the leave group fail with an `unknown_member_id`, which
happens when there is a problem between the client and the broker.  This
would cause the manager to exit abruptly and may be causing the rest of
the consumer group's processes to exit uncleanly.

Regardless of the error returned, we want to exit the consumer group
cleanly.