Separate Heartbeat from Manager
This moves the heartbeat functionality into a separate GenServer so that
even if the worker startup takes a long time, such as over high latency
connections, the broker won't remove the consumer from the group.

This seems to solve issues we've had where we were getting
:unknown_member_id errors back.

The heartbeat notifies the manager of a need for rebalance by exiting
with a reason of `:rebalance`. The Manager starts a rebalance upon
receiving that message.