Add KafkaEx.start_link_worker
This is a convenience function to start a worker outside the main
supervision tree.  Because it is outside the supervision tree, we may
start it with an arbitrary GenServer implementation, so this function
takes `server_impl` as an additional optional parameter.

I went back and cleaned up the kayrock integration tests to use this
function instead of a two-step process of building worker arguments and
then starting the client.