Add `dwell_on_empty` option for GenConsumer
I noticed that I was getting timeouts when testing consumer groups
locally.  I think this is due to the test consumer overwhelming my test
kafka cluster (running in docker).  I noticed that when we handle an
empty message set, we immediately poll again for new messages.  This is
probably what many, if not most, users want but I can also see where it
could be an issue for some users.  I added an option for GenConsumer
(which is passed through by ConsumerGroup options) to specify a number
of milliseconds to dwell if we receive an empty message set.  The
default is zero, which is equivalent to the existing behavior.