Fix default partitioner for kayrock
As reported on slack by @bartek - the default partitioner (really any
partitioner) needs to know how many partitions a topic has.  In the
Kayrock-based client, we don't automatically fetch metadata for every
topic.  So it becomes a problem if anyone tries to produce to a topic
that the client doesn't know about yet.

This should fix that by ensuring we have topic metadata for any topic
to which we produce.