Add describe_group/2,3 API in KafkaEx
This commit adds an implementation for Kafka's describe_groups API.
Given a consumer group name, we issue a Kayrock.DescribeGroups to the
broker controlling that group and return the response to the caller. In
case of errors, we rety the request up to 3 times by default.

Todo: This could be improved by adding DescribeGroups to
KafkaEx.Protocol and using that to intermediate the requests to the
brokers.